+++ app/api/auth.py
@@ -3,6 +3,8 @@ from app.services.auth_service import (
     get_google_auth_url,
     get_google_access_token,
     get_google_user_info,
+)
+from app.services.user_service import (
     authenticate_user,
     refresh_access_token,
     logout_user
